requirements.txt
fastapi
uvicorn
sqlalchemy
uvicorn[standard]
sqlmodel
asyncpg
pydantic
passlib[bcrypt]
python-jose
python-multipart
requests
pytest
httpx
pytest-asyncio
aiosqlite
celery
redis
pydantic-settings
greenlet